Added new method to alternate between soft and hard deletes for tei

+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/trackedentity/DefaultTrackedEntityInstanceService.java
@@ -585,7 +585,23 @@ public class DefaultTrackedEntityInstanceService
     public void deleteTrackedEntityInstance( TrackedEntityInstance instance )
     {
         attributeValueAuditService.deleteTrackedEntityAttributeValueAudits( instance );
-        trackedEntityInstanceStore.delete( instance );
+        deleteTrackedEntityInstance( instance, false );
+
+    }
+
+    @Override
+    public void deleteTrackedEntityInstance( TrackedEntityInstance instance, boolean forceDelete )
+    {
+        if ( forceDelete )
+        {
+            trackedEntityInstanceStore.delete( instance );
+        }
+        else
+        {
+            instance.setDeleted( true );
+            trackedEntityInstanceStore.update( instance );
+        }
+
     }
 
     @Override
